fix: ignore cases tagged as "NeuerFall = -1"
These records exist to document that cases were removed from the official RKI table, but we already detect their absense by comparing the table to the version from the day before. Currently we detect the missing records AND include the negative case numbers provided by the RKI as "NeuerFall = -1". This fixes itself the next day, but causes a small temporary negative spike. It should be safe for us to ignore these records.