build
The 'cases' table now contains two additional dimensions: age and gender. The 'sum' tables should not be affected by this change.

I noticed an issue where cases would not be removed from 'cases' if the new version of the rki table contained no more cases for that date in that region. This issue has been fixed, because it became much more relevant with two additional dimensions. This caused small changes to the 'sum' tables.